fix issue with FK's and add migrations to provider file

+++ migrations/2016_02_09_132439_create_permission_user_table.php
@@ -16,7 +16,7 @@ class CreatePermissionUserTable extends Migration
             $table->increments('id')->unsigned();
             $table->integer('permission_id')->unsigned()->index();
             $table->foreign('permission_id')->references('id')->on('permissions')->onDelete('cascade');
-            $table->integer('user_id')->unsigned()->index();
+            $table->unsignedBigInteger('user_id')->unsigned()->index();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->timestamps();
         });
